requirements.txt

streamlit
python-binance
pandas
matplotlib
